requirements.txt
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.15.4
pandas==0.23.4
python-dateutil==2.7.5
python-jose==3.0.1
python-jose-cryptodome==1.3.2
pytz==2018.7
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sklearn==0.0
Werkzeug==0.14.1